requirements.txt
Django==1.6.2
argparse==1.2.1
dj-database-url==0.2.2
django-multiselectfield==0.1.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
pycrypto==2.6.1
dj-static==0.0.6
django-tastypie==0.12.0
python-mimeparse==0.1.4